+++ src/Thumbnails/index.types.ts
@@ -2,6 +2,25 @@ import * as React from 'react'
 
 export type SpinoramaThumbnailsProps = {
 	/**
+	 * Determines how the slides are aligned.
+	 * Available options are:
+	 * - 'start': aligns the slides to the left (or top if `direction` is 'column');
+	 * - 'center': aligns the slides to center;
+	 * - 'end': aligns the slides to the right (or bottom if `direction` is 'column');
+	 *
+	 * @default 'start'
+	 */
+	justify?: 'start' | 'center' | 'end'
+
+	/**
+	 * Determines the direction of the slideshow.
+	 * When `true`, the slides will be displayed in reverse order.
+	 *
+	 * @default false
+	 */
+	reverse?: boolean
+
+	/**
 	 * The class name of the component.
 	 * Use this prop when you want to style the component with a className.
 	 */